requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
appdirs==1.4.4
appnope==0.1.2
astroid==2.5.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
beautifulsoup4==4.9.3
bowler==0.9.0
cachetools==4.2.1
Cerberus==1.3.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.1
colorama==0.4.4
decorator==4.4.2
dephell==0.8.3
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.4
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
dnspython==1.16.0
docker==5.0.0
dockerpty==0.4.1
docutils==0.17
fissix==20.8.0
Flask==1.1.1
google-api-core==1.26.3
google-auth==1.28.0
google-cloud==0.34.0
google-cloud-bigquery==2.13.1
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphviz==0.16
grpcio==1.36.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.1
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
m2r==0.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
moreorless==0.4.0
motor==2.1.0
multidict==5.1.0
numpy==1.18.1
packaging==20.9
pandas==1.0.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pip==20.0.2
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.6
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.7.4
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-gnupg==0.4.7
pytz==2019.3
pyzmq==22.0.3
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
setuptools==46.0.0
shellingham==1.4.0
six==1.13.0
soupsieve==2.2.1
tabulate==0.8.9
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
volatile==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.0
wheel==0.34.2
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
yaspin==1.5.0
